+++ tests/test_reusable_executor.py
@@ -733,70 +733,3 @@ class TestExecutorInitializer(ReusableExecutorMixin):
             assert x == 'uninitialized'
 
 
-def test_memory_leak_protection():
-
-    def leak_some_memory(size=int(1e6), delay=0.01):
-        if getattr(os, '__loky_leak', None) is None:
-            os.__loky_leak = []
-        os.__loky_leak.append(b"\x00" * size)
-
-        # Leave enough time for the memory leak detector to kick-in:
-        # by default the process does not check its memory usage
-        # more than once per second.
-        sleep(delay)
-
-        leaked_size = sum(len(buffer) for buffer in os.__loky_leak)
-        return os.getpid(), leaked_size
-
-    executor = get_reusable_executor(max_workers=1)
-
-    with pytest.warns(UserWarning, match='memory leak'):
-        futures = []
-        for i in range(300):
-            # Total run time should be 3s which is way over the 1s cooldown
-            # period between two consecutive memory checks in the worker.
-            futures.append(executor.submit(leak_some_memory))
-
-        results = [f.result() for f in futures]
-
-        # The pid of the worker has changed when restarting the worker
-        first_pid, last_pid = results[0][0], results[-1][0]
-        assert first_pid != last_pid
-
-        # The restart happened after 100 MB of leak over the default process
-        # size + what has leaked since the last memory check.
-        for _, leak_size in results:
-            assert leak_size / 1e6 < 250
-
-
-def test_reference_cycle_collection():
-    # make the parallel call create a reference cycle and make
-    # a weak reference to be able to track the garbage collected objects
-
-    class A:
-        def __init__(self, size=int(1e6)):
-            self.data = b"\x00" * size
-            self.a = self
-
-    def dummy_func(delay=0.001):
-        from loky import process_executor
-        process_executor._USE_PSUTIL = False
-        process_executor._GC_CHECK_DELAY = 0.1
-        if getattr(os, '__loky_cyclic_weakrefs', None) is None:
-            os.__loky_cyclic_weakrefs = []
-
-        a = A()
-        sleep(delay)
-        os.__loky_cyclic_weakrefs.append(weakref.ref(a))
-        return sum(1 for r in os.__loky_cyclic_weakrefs if r() is not None)
-
-    executor = get_reusable_executor(max_workers=1)
-
-    futures = []
-    for i in range(300):
-        # Total run time should be 3s which is way over the 1s cooldown
-        # period between two consecutive memory checks in the worker.
-        futures.append(executor.submit(dummy_func))
-
-    max_active_refs_count = max(f.result() for f in futures)
-    assert max_active_refs_count < 150
